requirements.txt
# SafeYatri Core Requirements
flask>=2.3.0
flask-socketio>=5.1.1
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
ultralytics>=8.0.0
python-dotenv>=1.0.0
email-validator>=1.1.3
Werkzeug>=2.3.0
Jinja2>=3.1.0
itsdangerous>=2.1.0
click>=8.1.0
MarkupSafe>=2.1.0
python-engineio>=4.7.0
python-socketio>=5.8.0
six>=1.16.0
pillow>=10.0.0
setuptools>=68.0.0
eventlet>=0.33.0
requests>=2.31.0
folium>=0.14.0

# Authentication & Security
pyjwt>=2.8.0
pyotp>=2.9.0
qrcode>=7.4.2
cryptography>=41.0.0
bcrypt>=4.0.0

# Monitoring & Logging
prometheus-client>=0.17.0
sentry-sdk>=1.32.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 